feat(v4.7.0): 幻影交付接线修复——配额/RLS/预热/N1 检测/cache_ttl 接入生产路径
接线（启用对应 feature 后生效，全部位于 #[cfg(feature)] 门控内）:
- pool.rs: QuotaEnforcer 接入 Pool（set_quota_enforcer + acquire_with_tenant/
  release_with_tenant 配额检查，PlMutex import cfg 放宽）
- query.rs: RlsPolicyEnhancer 接入 WHERE 构建链 + execute_with_cache
  接线 cache_ttl → L2Cache::get_or_load_query
- entity_graph.rs: N1QueryDetector 接入 BatchLoader（with_detector）
- process_l1_cache.rs: ProcessL1Cache::warmup 接入 CacheWarmer
- 接线集成测试：quota/warmup/cache_ttl/N1 全部通过

修复（接线验证暴露的存量问题）:
- l1_cache_bench.rs: put/get_or_load pk 参数 i64 类型不匹配（3 处）
- l1_cache.rs/process_l1_cache.rs: doctest 缺 mut、私有模块引用、async 上下文

验证: cargo check/clippy(-D warnings)/fmt 全绿；feature 测试 61 passed；
门禁 15 PHANTOM-1 37→34；全 workspace 测试仅 1 个存量 flaky（sz-orm-storage 环境相关）

+++ packages/sz-orm-core/benches/l1_cache_bench.rs
@@ -11,7 +11,8 @@ fn bench_l1_hit(c: &mut Criterion) {
     c.bench_function("l1_cache_hit", |b| {
         let mut cache: L1Cache<i32> = L1Cache::new(1000);
         for i in 0..1000 {
-            cache.put(i, Arc::new(i));
+            // put 的 key 参数为 i64（l1_cache.rs:126），value 需要 Arc<i32>
+            cache.put(i, Arc::new(i as i32));
         }
         b.iter(|| {
             for i in 0..1000 {
@@ -37,7 +38,7 @@ fn bench_l1_put(c: &mut Criterion) {
         b.iter(|| {
             let mut cache: L1Cache<i32> = L1Cache::new(1000);
             for i in 0..100 {
-                cache.put(i, Arc::new(i));
+                cache.put(i, Arc::new(i as i32));
             }
             black_box(&cache);
         })
@@ -49,7 +50,8 @@ fn bench_l1_l2_db_db_load(c: &mut Criterion) {
         let mut coord: L1L2Coordinator<i32> = L1L2Coordinator::new(1000);
         b.iter(|| {
             for i in 0..100 {
-                black_box(coord.get_or_load("t", i, || Some(i)));
+                // get_or_load 的 pk 参数为 i64（l1_cache.rs:245），循环变量 i 默认 i32
+                black_box(coord.get_or_load("t", i as i64, || Some(i)));
             }
         })
     });
@@ -59,11 +61,11 @@ fn bench_l1_l2_db_l1_hit(c: &mut Criterion) {
     c.bench_function("l1_l2_db_l1_hit", |b| {
         let mut coord: L1L2Coordinator<i32> = L1L2Coordinator::new(1000);
         for i in 0..1000 {
-            coord.get_or_load("t", i, || Some(i));
+            coord.get_or_load("t", i as i64, || Some(i));
         }
         b.iter(|| {
             for i in 0..100 {
-                black_box(coord.get_or_load("t", i, || Some(i)));
+                black_box(coord.get_or_load("t", i as i64, || Some(i)));
             }
         })
     });
